Change version type in AccountHistoryContext from Integer to BigInteger

+++ src/main/java/uk/gov/hmcts/opal/service/opal/history/core/AccountHistoryContext.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.opal.service.opal.history.core;
 
+import java.math.BigInteger;
 import lombok.Getter;
 
 @Getter
@@ -9,19 +10,19 @@ public class AccountHistoryContext {
 
     private final Long accountId;
 
-    private final Integer version;
+    private final BigInteger version;
 
     public AccountHistoryContext(AccountHistoryType accountType, Long accountId) {
         this(accountType, accountId, null);
     }
 
-    public AccountHistoryContext(AccountHistoryType accountType, Long accountId, Integer version) {
+    public AccountHistoryContext(AccountHistoryType accountType, Long accountId, BigInteger version) {
         this.accountType = accountType;
         this.accountId = accountId;
         this.version = version;
     }
 
-    public AccountHistoryContext withVersion(Integer version) {
+    public AccountHistoryContext withVersion(BigInteger version) {
         return new AccountHistoryContext(accountType, accountId, version);
     }
 }
